Refactored build to use make.go
This commit creates a file called make.go which takes care of building
the project. Makefiles just call this to do the build. This approach is
required to support cross-platform builds without managing different
build toolchains

  go run make.go -> Will do a full compile
  go run make.go --install -> Will do a install to the default prefix
  go run make.go --install --prefix /tmp -> Installs to /tmp
  go run make.go --test -> Runs the tests

On Make supported machines, this can be executed by

  make
  make install
  prefix=/tmp make install
  make test

This also combines, gauge-java and common to the gauge repository. No
more submodules and different repos