[java] compiling all files under src
Before it was compiling only the files which are under "src/test/java".
This is very restricting and not all of them will have project structure
like that. So to support any types of package structure, we compile all
the files under src.

TODO: Make this a configurable property so that incase of large number
of files, we can compile just the test directory